+++ package.json
@@ -8,6 +8,7 @@
     "airtable": "^0.8.1",
     "dotenv-safe": "^8.2.0",
     "esm": "^3.2.25",
+    "express": "^4.17.1",
     "moment": "^2.24.0",
     "node-fetch": "^2.6.0"
   },
